internal/scte35: don't refactor, reimplement scte35 package
This is a huge WIP patch which should never have come out like this,
but we made it all while live streaming so it will have to do for now.
All external dependencies are removed, and there are far fewer types
to juggle. Progress so far is we're almost able to encode SpliceInfo.