rtp: implement RTP sessions and packet transmission
This is not meant to be performant or robust to any flaky networks.
Instead it lets Go's test runner generate and consume streams by
starting a session on the local system, transmit some packets to a
fake player, then check if the packets are formed ok. Now we finally
have a way to mess with our packets and headers and get fast feedback
without involving some other program like ffmpeg or VLC to do
streaming for us!