sdp: add basic media description parsing
Not fully tested; need to test e.g. port count. The Media struct is
probably not complete yet, either, as we may need to add connection
info (the data from the "c=" line from the SDP text). For now we parse
the example SDP text from the RFC, so worth a commit!