sdp: only test SDP duration parsing
All other durations are handed over to time.ParseDuration, so we
can let the standard library take care of it. Since this part of
SDP is rarely - if ever - used in the wild, I'd rather have less
code.

While here don't bother with subtests; letting it all run one after
the other is just as clear and less code. We're not doing parallel
testing either.