cmd/autx: correctly copy bytes out to the network
We weren't really copying them before, we were just grabbing whatever
we had in our buffer which was constantly being overwritten!

This patch also correctly specifies and sets the payload size to
always fits a UDP packet to fit in the typical network MTU of 1400-ish
bytes. The sample rate and transmission interval are modified
accordingly with guidance from RFC 3551.