m3u8: clarify purpose of some sample playlists
Big Buck Bunny is the classic test sample for a lot of video dev work.
tos.m3u8 is here because it is a valid playlist containing lots of
comments and empty lines. Unusual, but still a valid playlist; hoping
it will help catch bugs in our parser's handling of comments and
whitespace.